+++ src/components/atoms/Category/Category.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react'
+
 import { Category } from '.'
 import { Tag } from '../Tag'
