+++ src/components/atoms/Category/index.tsx
@@ -1,11 +1,9 @@
 import styled from '@emotion/styled'
+
 import { theme } from '@styles/theme'
-import { RoundButton } from '../RoundButton'
-import { Text } from '../Text'
-import Svg from '../Svg'
-import Plus from '@assets/icons/Plus'
-import Gift from '@assets/icons/Gift'
-import { SubCategoryName } from '@utils/apis/user/UserType'
+import { CategoryPropsType } from '@models/components/atoms/Category'
+import CategoryHeader from './CategoryHeader'
+import TagListItem from './TagListItem'
 
 /**
  * @param categoryName 카테고리명을 나타냄
@@ -15,26 +13,6 @@ import { SubCategoryName } from '@utils/apis/user/UserType'
  * @param onPlusButtonClick + / 선물 버튼 클릭 시 발생하는 이벤트를 넘겨주는 함수임
  */
 
-export type CategoryNameType =
-  | 'MAKEUP'
-  | 'FRAGRANCE'
-  | 'CLOTHES'
-  | 'FASHION_PRODUCT'
-  | 'DIGITAL_PRODUCT'
-  | 'BAG'
-  | 'ACCESSORY'
-  | 'COOKING'
-  | 'EXERCISE'
-
-type CategoryPropsType = {
-  categoryName: SubCategoryName
-  isFriend: boolean
-  children: React.ReactNode[]
-  allCategoryAnswered?: boolean
-  onPlusButtonClick?: () => void
-  onPresentButtonClick?: () => void
-}
-
 export const Category = ({
   categoryName,
   isFriend = false,
@@ -43,45 +21,16 @@ export const Category = ({
   onPlusButtonClick,
   onPresentButtonClick,
 }: CategoryPropsType) => {
-  const renderIcon = () => {
-    if (isFriend) {
-      return <Svg children={<Gift />} onClick={onPresentButtonClick} />
-    } else {
-      if (!allCategoryAnswered) {
-        return <Svg children={<Plus />} onClick={onPlusButtonClick} />
-      } else {
-        return null
-      }
-    }
-  }
-
   return (
     <Wrapper>
-      <Heading>
-        <Text
-          children={categoryName}
-          typo="Headline_16"
-          color="white"
-          as="div"
-        />
-        {isFriend || !allCategoryAnswered ? (
-          <RoundButton
-            variant="circle"
-            children={
-              <Svg
-                children={renderIcon()}
-                style={{ display: 'flex', alignItems: 'center' }}
-                height="fit-content"
-              />
-            }
-          />
-        ) : null}
-      </Heading>
-      <TagWrapper>
-        {children.map((child, index) =>
-          child ? <TagItem key={index}>{child}</TagItem> : null,
-        )}
-      </TagWrapper>
+      <CategoryHeader
+        categoryName={categoryName}
+        isFriend={isFriend}
+        allCategoryAnswered={allCategoryAnswered}
+        onPlusButtonClick={onPlusButtonClick}
+        onPresentButtonClick={onPresentButtonClick}
+      />
+      <TagListItem children={children} />
     </Wrapper>
   )
 }
@@ -93,21 +42,3 @@ const Wrapper = styled.div`
   padding: 16px 18px 20px 18px;
   background-color: ${theme.palette.gray_800};
 `
-
-const Heading = styled.div`
-  width: 100%;
-  height: 24px;
-  margin-bottom: 16px;
-  display: flex;
-  justify-content: space-between;
-`
-
-const TagWrapper = styled.div`
-  width: 100%;
-  height: fit-content;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 12px;
-`
-
-const TagItem = styled.div``
