+++ src/models/favor.ts
@@ -1,0 +1,10 @@
+export type SubCategoryNameType =
+  | '메이크업'
+  | '프레그런스'
+  | '의류'
+  | '패션소품'
+  | '디지털소품'
+  | '가방'
+  | '액세사리'
+  | '요리'
+  | '운동'
